requirements.txt
apturl==0.5.2
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
distro==1.7.0
distro-info===1.1build1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.4
duplicity==0.8.21
exceptiongroup==1.0.0rc9
facebook-sdk==3.1.0
Faker==15.0.0
fasteners==0.14.1
flatbuffers==22.10.26
future==0.18.2
GDAL==3.4.1
google-auth==2.14.1
h11==0.14.0
httplib2==0.20.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imageio==2.22.3
importlib-metadata==4.6.4
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.2.1
msgpack==1.0.4
netifaces==0.11.0
networkx==2.8.8
nudepy==0.5.1
numpy==1.23.4
oauthlib==3.2.0
olefile==0.46
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
outcome==1.2.0
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
progressbar2==4.2.0
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydload==1.0.9
PyGObject==3.42.1
PyJWT==2.5.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-utils==3.4.5
pytz==2022.2.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.3.4
reportlab==3.6.8
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.3
scipy==1.9.3
SecretStorage==3.3.1
selenium==4.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
sympy==1.11.1
systemd-python==234
texttable==1.6.4
tifffile==2022.10.10
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.4
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.12
usb-creator==0.3.7
wadllib==1.3.6
websocket-client==1.2.3
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
